Do a little test refactoring.
